Add files via upload
In this project we use many this which gives us a correct visualisation of the data like pandas-profiling and heatmap. Because of these there is lot of data stored in this file by which we can visualise data very clearly and make a model for that. In this project we actually find a algorithm on which our data give a good prediction. For that we use here RandomForestClassifier and Naive Bayes algorithm on makiing model on the basis of this both algorithm we can conclude that our Naive Bayes gives good prediction on our test data. So we select this model for our prediction that which customer give money back.